% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dig_root.R
\name{dig_root}
\alias{dig_root}
\title{Find the digital root of an integer}
\usage{
dig_root(x)
}
\arguments{
\item{x}{an integer}
}
\value{
the digital root of the integer
}
\description{
This function will calculate the sum of all the single-length
integers [0-9] in a given integer recursively until the result is a
single-length integer
}
\examples{
dig_root(77) # 7 + 7 = 14; 1 + 4 = 5;

# Get the multiplication table of digital roots
sapply(1:9, function(i) dig_root(i * 1:9))
}
\references{
From Thomas Morril on Twitter:

\url{https://twitter.com/tsmorrill/status/1028108510746226688}
\url{https://twitter.com/tsmorrill/status/1028142110342504448}
}
\seealso{
\code{\link[=dig_sum]{dig_sum()}}
}
