% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dig_sum.R
\name{dig_sum}
\alias{dig_sum}
\title{Sum the digits of a number}
\usage{
dig_sum(x)
}
\arguments{
\item{x}{an integer}
}
\value{
the sum of all the individual digits of that integer
}
\description{
Sum the digits of a number
}
\examples{
dig_sum(10)  # 1
dig_sum(123) # 6
}
\seealso{
\code{\link[=dig_root]{dig_root()}}
}
